+++ packages/brick_offline_first_build/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 * Apply standardized lints
+* Support using `OfflineFirstSerdes` as a unique field in SQLite
 
 ## 3.0.1