+++ packages/brick_sqlite/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Apply standardized lints
 * Upgrade minimum Dart to 2.18
+* Add extra documentation to the `@Sqlite(unique:)` annotation
 
 ## 3.0.1